+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Avoid including Colossal.Mathematics dll
 - Fix an issue where the utilization of the Multi-Level Parking Building was incorrect
+- Fix delay where policy updates would not apply until the start of the next calculation cycle
 
 ## Version 1.0.1